+++ lal/utils/conversions.hpp
@@ -55,7 +55,7 @@ namespace utils {
  * @param n Number of nodes of the tree.
  * @return Returns the tree built with @e L.
  */
-graphs::utree level_sequence_to_tree(const std::vector<uint32_t>& L, uint32_t n);
+graphs::utree level_sequence_to_tree(const std::vector<uint64_t>& L, uint64_t n);
 
 /*
  * @brief Converts a linear sequence of a tree to a graph structure.
@@ -68,7 +68,7 @@ graphs::utree level_sequence_to_tree(const std::vector<uint32_t>& L, uint32_t n)
  * @param n Number of nodes of the tree.
  * @return Returns the tree built with @e L.
  */
-graphs::utree linear_sequence_to_tree(const std::vector<uint32_t>& L, uint32_t n);
+graphs::utree linear_sequence_to_tree(const std::vector<uint64_t>& L, uint64_t n);
 
 /*
  * @brief Converts the Prüfer sequence of a labelled tree into a tree structure.
@@ -81,7 +81,7 @@ graphs::utree linear_sequence_to_tree(const std::vector<uint32_t>& L, uint32_t n
  * @param n Number of nodes of the tree.
  * @return Returns the tree built with @e L.
  */
-graphs::utree Prufer_sequence_to_tree(const std::vector<uint32_t>& S, uint32_t n);
+graphs::utree Prufer_sequence_to_tree(const std::vector<uint64_t>& S, uint64_t n);
 
 } // -- namespace utils
 } // -- namespace lal
